test(useTableDragSelect): add test
